+++ core/config.py
@@ -50,6 +50,7 @@ LONG_LINE = '-' * 50
 
 SKALE_NODE_UI_PORT = 3007
 SKALE_NODE_UI_LOCALHOST = 'http://0.0.0.0'
+DEFAULT_URL_SCHEME = 'http://'
 
 SKALE_PROJECT_PATH = os.path.join('/skale', 'skale-node')
 UNINSTALL_SCRIPT = os.path.join(SKALE_PROJECT_PATH, 'installation', 'uninstall.sh')